fix(#3002): memory management
