+++ eo-runtime/src/main/eo/org/eolang/cage.eo
@@ -32,4 +32,4 @@
 [] > cage
   # Make new `cage` for an object. After application this `new` object starts behave
   # like `it` object.
-  [it] > new /?
+   [it] > new /?
