+++ eo-runtime/src/main/eo/org/eolang/malloc.eo
@@ -26,14 +26,26 @@
 +rt jvm org.eolang:eo-runtime:0.0.0
 +version 0.0.0
 
-# Dynamic memory allocation manager.
+# Allocates a block in memory of `size` bytes.
+# You can use it like:
+# ```
+# (malloc 8).pointer > p
+# seq
+#   *
+#     p.write 42 # write 8 bytes integer
+#     p.read     # read from memory, 42 will be returned
+#     p.free     # free allocated block
+# ```
+# Clearing the block is optional and is up to programmer.
 [size] > malloc
-  # Allocates a block in memory of `size` bytes, and returns a pointer to it.
-  [] > @ /pointer
+  @.pointer > pointer
+
+  # Allocates a memory block in RAM and returns pointer to it.
+  [] > @ /memory-block-pointer
 
   # Pointer to allocated block in memory.
   # Here `id` is identifier of pointer, `size` is length of the block.
-  [id size] > pointer
+  [id size] > memory-block-pointer
     $ > pointer
 
     # Read `bytes` from the block in memory by the pointer.
