+++ eo-runtime/src/main/java/EOorg/EOeolang/EOmalloc$EOmemory_block_pointer$EOwrite.java
@@ -28,9 +28,11 @@
 
 package EOorg.EOeolang;
 
+import org.eolang.AtVoid;
 import org.eolang.Atom;
 import org.eolang.Attr;
 import org.eolang.Data;
+import org.eolang.Dataized;
 import org.eolang.Heaps;
 import org.eolang.Param;
 import org.eolang.PhDefault;
@@ -39,27 +41,29 @@ import org.eolang.Versionized;
 import org.eolang.XmirObject;
 
 /**
- * Malloc.pointer.read object.
+ * Malloc.pointer.write object.
  * @since 0.36.0
  * @checkstyle TypeNameCheck (5 lines)
  */
 @Versionized
-@XmirObject(oname = "malloc.pointer.read")
-final class EOmalloc$EOpointer$EOread extends PhDefault implements Atom {
+@XmirObject(oname = "malloc.pointer.write")
+final class EOmalloc$EOmemory_block_pointer$EOwrite extends PhDefault implements Atom {
     /**
      * Ctor.
      * @param sigma Sigma
      */
-    EOmalloc$EOpointer$EOread(final Phi sigma) {
+    EOmalloc$EOmemory_block_pointer$EOwrite(final Phi sigma) {
         super(sigma);
+        this.add("data", new AtVoid("data"));
     }
 
     @Override
     public Phi lambda() throws Exception {
-        return new Data.ToPhi(
-            Heaps.INSTANCE.get().read(
-                new Param(this.take(Attr.RHO), "id").strong(Long.class).intValue()
-            )
+        final Phi rho = this.take(Attr.RHO);
+        Heaps.INSTANCE.get().write(
+            new Param(this.take(Attr.RHO), "id").strong(Long.class).intValue(),
+            new Dataized(this.take("data")).take()
         );
+        return new Data.ToPhi(true);
     }
 }
