+++ eo-runtime/src/test/eo/org/eolang/malloc-tests.eo
@@ -66,7 +66,7 @@
 # Test.
 [] > malloc-pointer-cant-write-more-than-allocated
   (malloc 1).pointer > p
-  try
+  try > @
     seq
       *
         p.write 42
